+++ hal/port/unix/lt_port_raspberrypi_wiringpi.c
@@ -25,8 +25,8 @@
 // File descriptor, used in init and deinit
 int fd = 0;
 
-// #define NDEBUG
-#ifdef NDEBUG
+// #define LIBT_DEBUG
+#ifdef LIBT_DEBUG
 #define LOG_OUT(f_, ...) printf("[TCP] " f_, ##__VA_ARGS__)
 #define LOG_ERR(f_, ...) fprintf(stderr, "ERROR: " f_, ##__VA_ARGS__)
 #define LOG_U8_ARRAY(arr, len)      \